passage de Long a int pour les id questions et topics dans le back

+++ jang/src/main/java/com/bdr/jang/repository/QuestionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 public interface QuestionRepository
-        extends JpaRepository<Question, Long>, JpaSpecificationExecutor<Question> {
+        extends JpaRepository<Question, Integer>, JpaSpecificationExecutor<Question> {
 
 
 }
